+++ ui/src/flux/components/FetchFields.tsx
@@ -2,14 +2,14 @@
 import {PureComponent} from 'react'
 
 // Utils
-import {fieldsByMeasurement as fetchFieldsByMeasurementAsync} from 'src/shared/apis/flux/metaQueries'
-import {parseFieldsByMeasurements} from 'src/shared/parsing/flux/values'
+import {fetchFieldsByMeasurement} from 'src/shared/apis/flux/metaQueries'
 
 // Types
-import {Source, RemoteDataState} from 'src/types'
+import {Source, RemoteDataState, TimeRange} from 'src/types'
 
 interface Props {
   source: Source
+  timeRange: TimeRange
   bucket: string
   children: (fields, fieldsByMeasurement, fieldsLoading) => JSX.Element
 }
@@ -43,13 +43,13 @@ class FetchFields extends PureComponent<Props, State> {
   }
 
   private async fetchFields() {
-    const {source, bucket} = this.props
+    const {source, timeRange, bucket} = this.props
     this.setState({loading: RemoteDataState.Loading})
     try {
-      const fieldsResults = await fetchFieldsByMeasurementAsync(source, bucket)
-
-      const {fields, fieldsByMeasurements} = parseFieldsByMeasurements(
-        fieldsResults
+      const {fields, fieldsByMeasurements} = await fetchFieldsByMeasurement(
+        source,
+        timeRange,
+        bucket
       )
 
       this.setState({