+++ ui/src/flux/components/FetchTagKeys.tsx
@@ -2,14 +2,14 @@
 import {PureComponent} from 'react'
 
 // Utils
-import parseValuesColumn from 'src/shared/parsing/flux/values'
-import {tagKeys as fetchTagKeysAsync} from 'src/shared/apis/flux/metaQueries'
+import {fetchTagKeys} from 'src/shared/apis/flux/metaQueries'
 
 // Types
-import {Source, RemoteDataState} from 'src/types'
+import {Source, RemoteDataState, TimeRange} from 'src/types'
 
 interface Props {
   source: Source
+  timeRange: TimeRange
   bucket: string
   children: (tagKeys, tagsLoading) => JSX.Element
 }
@@ -29,21 +29,18 @@ class FetchTagKeys extends PureComponent<Props, State> {
   }
 
   public componentDidMount() {
-    this.fetchTagKeys()
+    this.fetchData()
   }
 
   public render() {
     return this.props.children(this.state.tagKeys, this.state.loading)
   }
 
-  private async fetchTagKeys() {
-    const {source, bucket} = this.props
+  private async fetchData() {
+    const {source, timeRange, bucket} = this.props
     this.setState({loading: RemoteDataState.Loading})
     try {
-      const tagKeysResults = await fetchTagKeysAsync(source, bucket, [])
-
-      const tagKeys = parseValuesColumn(tagKeysResults)
-
+      const tagKeys = await fetchTagKeys(source, timeRange, bucket)
       this.setState({tagKeys, loading: RemoteDataState.Done})
     } catch (error) {
       this.setState({loading: RemoteDataState.Error})