+++ ui/src/shared/components/TimeMachine/fluxQueryBuilder/util/generateFlux.ts
@@ -1,14 +1,6 @@
 import fluxString from 'src/flux/helpers/fluxString'
-import {TimeRange} from 'src/types'
 import {BuilderTagsType, QueryBuilderState} from '../types'
 import {AGG_WINDOW_AUTO, FUNCTIONS} from './constants'
-
-export function formatTimeRangeArguments(timeRange: TimeRange): string {
-  const start = timeRange.lowerFlux ? timeRange.lowerFlux : timeRange.lower
-  return timeRange.upper
-    ? `start: ${start}, stop: ${timeRange.upper}`
-    : `start: ${start}`
-}
 
 export function tagToFlux(tag: BuilderTagsType) {
   return tag.tagValues