+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Features
 
+1. [#5852](https://github.com/influxdata/chronograf/pull/5852): Add Flux Query Builder.
+1. [#5858](https://github.com/influxdata/chronograf/pull/5858): Use time range in flux Schema Explorer.
+
 ### Bug Fixes
 
 1. [#5862](https://github.com/influxdata/chronograf/pull/5862): Respect BASE_PATH when serving API docs.